requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-allauth==0.57.0
dj-rest-auth==5.0.2
python-decouple==3.8
django-filter==23.5
django-extensions==3.2.3
gunicorn==21.2.0
